Add additional details to voting form. Consolidate date functions
