+++ src/constants/gauge-controller.ts
@@ -1,1 +1,3 @@
-export const WEIGHT_VOTE_DELAY = 10 * 86400;
+import { oneDayInMs } from "@/composables/useTime";
+
+export const WEIGHT_VOTE_DELAY = 10 * oneDayInMs;
