--- src/lib/utils/date.ts
@@ -1,14 +1,0 @@
-export function toUtcTime(date: Date) {
-  return Date.UTC(
-    date.getUTCFullYear(),
-    date.getUTCMonth(),
-    date.getUTCDate(),
-    date.getUTCHours(),
-    date.getUTCMinutes(),
-    date.getUTCSeconds()
-  );
-}
-
-export function unixToJsTime(time: number) {
-  return time * 1000;
-}
